unify notcurses-input+notcurses-keyplot #455
